requirements.txt
requests==2.31.0
Pillow==10.1.0
fastapi==0.105.0
python-dotenv==1.0.0
uvicorn==0.24.0.post1